requirements.txt
langgraph
langchain
langchain_community
faiss-cpu
pandas
numpy
scikit-learn
sentence-transformers
jupyter
ipykernel
groq
langchain_huggingface
python-dotenv
zyndai-agent
streamlit